requirements.txt
alembic==1.15.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector==2.2.9
python-dotenv==1.0.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
openai==1.79.0
